Move ssNN and cvNN feature stuff into AllFeatures
This was previously living with the tables, which doesn't make much
sense
